further use of KualiDecimal in the XML files

+++ src/main/java/edu/cornell/kfs/sys/xmladapters/KualiDecimalXmlAdapter.java
@@ -1,5 +1,7 @@
 package edu.cornell.kfs.sys.xmladapters;
 
+import java.math.BigDecimal;
+
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
 import org.kuali.rice.core.api.util.type.KualiDecimal;
@@ -8,7 +10,7 @@ public class KualiDecimalXmlAdapter extends XmlAdapter<String, KualiDecimal> {
 
     @Override
     public KualiDecimal unmarshal(String v) throws Exception {
-        Double doubleValue = new Double(v);
+        BigDecimal doubleValue = new BigDecimal(v);
         return new KualiDecimal(doubleValue);
     }
